refactor: 메서드 이름 변경
